requirements.txt
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
pillow==11.3.0
drf-spectacular==0.28.0
django-debug-toolbar==6.0.0
pytest==8.4.2
pytest-django==4.11.1
stripe==12.5.1
python-dotenv==1.1.1
